Set fromEntry to nil after discardNonCommittedItems.
With the indirection CRWSessionEntry->NavigationItem,
discardNonCommittedItems can invalidate fromEntry.
Set if to nil if it is invalid.

This is a tentative fix by the sheriff to make the tree green.
Please review this CL and revert it later if there is a better solution.

BUG=691492
TBR=eugenebut@chromium.org, kkhorimoto@chromium.org,

Review-Url: https://codereview.chromium.org/2690033002
Cr-Original-Commit-Position: refs/heads/master@{#449941}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f58024526f7dfa06703d876beac569242299357