Make test-specific Xcode version explicit
The top-level Xcode version was previously used to determine what version to use for compilation, and was used as a fallback when the test failed to specify an Xcode version. However, with the advent of hermetic build, top-level Xcode version is no longer used to determine what version is used for compilation.

BUG=686304

Review-Url: https://codereview.chromium.org/2687873003
Cr-Original-Commit-Position: refs/heads/master@{#450153}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cc60b74fd812c97f3495ec777fd0260649b7cb71