Fixed navigation for displayed SSL interstitials.
Navigation did not work for interstitials for 2 reasons which are fixed
in this CL:

1.) |WebInterstitialImpl::DontProceed| should not reload.
Reloading stops pending navigation and reloads last committed item,
which is not always correct target item. On other platforms
InterstitialPageImpl::DontProceed does not reload.

2.) |clearTransientContentView| removes pending item, so it should be
called before pending navigation is initiated. So next
|clearTransientContentView| call will be no-op.

This CL changes the behavior, so when user taps "BACK TO SAFETY" the page
is not reloaded. This is fine, but in order to update omnibox URL, tab
should subscribe to webStateDidDismissInterstitial: callback and update
the toolbar.

BUG=677193,677190,688009

Review-Url: https://codereview.chromium.org/2687343003
Cr-Original-Commit-Position: refs/heads/master@{#450220}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b2517f141263ee8913024a3a96cd909423ad7214