Use IOSImageDataFetcherWrapper for favicon
Use IOSImageDataFetcher instead of ImageDataFetcher for downloading the favicon
on iOS. This allows the download to be done in the favicon driver instead of the
WebState.
The favicon needs the http response code for optimization (if the favicon does
not exist we try to download it only once).

BUG=683918

Review-Url: https://codereview.chromium.org/2677993002
Cr-Original-Commit-Position: refs/heads/master@{#450317}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 911c2949fcb7a8362449e8af1d60bbb93d3cfab5