[iOS Reading List] Add distilling WKWebView to the view hierarchy.
Some pages (e.g. uk.businessinsider.com) are not rendered correctly
unless the view is added in the view hierarchy.
Add temporarily the WKWebView to the view hierarchy during distillation.
Move it out ouf the screen to make it invisible to the user.

BUG=685583

Review-Url: https://codereview.chromium.org/2695313004
Cr-Original-Commit-Position: refs/heads/master@{#451643}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 47b54ff088d9c52ece1c28b06e647e64ec8014d7