Extend Physical Web iOS auto-enable behavior beyond first launch.
On Chrome start, we check whether the device state permits us to
auto-enable the Physical Web feature to perform foreground scanning for
nearby Bluetooth LE devices broadcasting URLs.

Previously, if the auto-enable conditions were not met, the Physical Web
preference would be set to the explicitly-disabled state. With this change,
the preference is left in an onboarding state and Chrome will check again
the next time Chrome is started.

BUG=695184

Review-Url: https://codereview.chromium.org/2704343005
Cr-Original-Commit-Position: refs/heads/master@{#452571}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e3b410bacaa28cde0c2be4c2eba96451c6ca72a