[tab_container] Hack-in fix for broken constraint on animation
Custom view controller transition animations must be designed
so as not to break constraints on the presented/dismissed VC's view.

This CL addresses a broken constraint due to the zoom animator.
The zoom animator will resize the VC's view to a small rect.
However, that rect is smaller in height than the sum of the
height constraints on the subviews of the tab_container.

This CL simply reduces the subview height constant. Since the
transition animation will likely change, there is no need to modify
the zoom animator at this time.

BUG=

Review-Url: https://codereview.chromium.org/2717133002
Cr-Original-Commit-Position: refs/heads/master@{#453249}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1bcc418f288f8b84643f0f8b34cbe7eb0fd548bd