Test Scheme of page before adding a ReadingListActivity.
If user has a physical keyboard, it is possible to show the share menu
on non http[s] pages, which are not supported in the Chrome iOS Reading List.
This CL checks the scheme of the page before adding the "Read Later" entry.

Review-Url: https://codereview.chromium.org/2722293003
Cr-Original-Commit-Position: refs/heads/master@{#454223}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 639e229373e9cb69c1259648cedc70a5841173e6