[Remote suggestions] Move all decisions to fetch to the scheduler
This change will clarify responsibilities between the scheduler and the
provider. On top of that, it eliminates the need to fetch snippets upon
creating the provider (which is dangerous w.r.t. fetches in unexpected
situations).

This change needs the new types of notifications from the provider to
the scheduler:
 - suggestions have been cleared (we should fetch on any later trigger),
 - history has been cleared (we should NOT fetch for a while).

Thus, the provider is now given a pointer to the scheduler interface.
(the interface now mixes internal and external signals but I think it is
this way easier to understand then with two separate interfaces).

BUG=694324

Review-Url: https://codereview.chromium.org/2702223004
Cr-Original-Commit-Position: refs/heads/master@{#454240}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6dd692b945fc8599a4c7db22854d19c9b27d0d4c