Move the notion of current Tab from TabModel to WebStateList.
Add methods to track and set the currently active WebState in the
WebStateList (referenced by index) and corresponding events when
the active WebState is changed (explicitly or automatically after
detaching a WebState).

Add new WebStateListObserver (and WebStateListObserving when the
observer need to keep a __weak pointer to an Objective-C class)
that react to the event send when the active WebState changes.

Add a -loadFinished property to Tab to avoid accessing the private
//ios/web private API in new code.

BUG=687207

Review-Url: https://codereview.chromium.org/2703333006
Cr-Original-Commit-Position: refs/heads/master@{#454329}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2f7781ce46951904a2615fbda44d2a9f4ced23ca