Convert TabModelTest to use the public API of TabModel.
Give a unique window name to all the inserted Tabs and disable web
usage on the TabModel. With those changes it is possible to restrict
TabModelTest to use the public API of TabModel.

Using the public API makes it easier to refactor the code (as there
is no dependency on the implementation or hole into it from the test).

BUG=687207

Review-Url: https://codereview.chromium.org/2715753002
Cr-Original-Commit-Position: refs/heads/master@{#454345}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dd0cb3cdf573d6ea0021f2962ef8116c9e80857