Created web::UserAgentType.
Previously, NavigationItems contained a bool denoting whether a
desktop user agent is used.  However, the user agent as a concept does
not apply to app-specific URLs, which are not loaded using HTTP
requests.  This CL introduces a UserAgentType enum, which contains the
values NONE, MOBILE, and DESKTOP in order to differentiate between
app-specific navigations and those using mobile/desktop user agents.

This should assist with logic pertaining to whether to expose the
option to use the desktop/mobile user agent, as app-specific URLs are
handled by the enum type.

BUG=none

Review-Url: https://codereview.chromium.org/2705293014
Cr-Original-Commit-Position: refs/heads/master@{#454426}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d79e11f6225265fcf1d5884b969796a0377dee8a