Send IDC_HIDE_PAGE_INFO to close PageInfoView when reloading the page.
In M-58+, BVC must be informed if the Page Info controller is dismissed.
This is caused by the check added in https://codereview.chromium.org/2672023002
that prevent displaying twice the page info.
Use the command to dismiss the popup instead of dismissing it directly.

BUG=697470

Review-Url: https://codereview.chromium.org/2723903005
Cr-Original-Commit-Position: refs/heads/master@{#455067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 61f85a6f747a80528881e5e3fa4d12d20325dc4b