Reenable HistoryEntryInserterTest.AddSections.
The test added history entries at 1am the day of the test, the day before and 2 days before and expects the three to appear in different sections.
With DST, 48hours before was still the same day as 24h ago.
This test would have turned green tomorrow, but now that it is moved to middays, we shouldn't run in the issue again next time.

BUG=700814
R=gambard@chromium.org

Review-Url: https://codereview.chromium.org/2747713002
Cr-Original-Commit-Position: refs/heads/master@{#456356}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 88543c5651a7509cf1d39370504a597081b240e9