Move resized_avatar_cache and account_control_item to authentication
In order to implement crbug.com/661794, a sign-in promo view is needed in
authentication folder which needs to be used by the settings collection view.

Right now account_control_item.mm and resized_avatar_cache.mm which are inside
settings (cells and utils) depend on authentication. So to be able to create the
sign-in promo view, those 2 files have to be moved into authentication.
Then the sign-in promo view can be used by the settings and some other views
(like bookmarks view).

BUG= 661794

Review-Url: https://codereview.chromium.org/2747893002
Cr-Original-Commit-Position: refs/heads/master@{#456679}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84c1f04f3195eaadc2cc1b8a07a7d73acbcf05b2