[ios] Remove cleanup code in Tab -close method.
Remove uses of private ios/web API in Tab -close method as the
method TabModel -didCloseTab: does not access the transient or
pending items (only access the navigation history while saving
the session) so there is no need to discard transient items.

Furthermore, there is no access to the request context in the
TabModel -didCloseTab: so there is no need to terminate all the
network activity in Tab -close (it will be stopped when the
CRWWebSessionController will be closed as part of WebStateImpl
destructor).

BUG=620465,546222

Review-Url: https://codereview.chromium.org/2747833004
Cr-Original-Commit-Position: refs/heads/master@{#456692}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d2131a095d1d947cd28a7cd4effc913238827440