Removed CRWWebDelagate callbacks for load cancellation.
When load is cancelled Tab already receives webStateDidStopLoading:.
In that callback it is safe to call |enableFullScreen| because the page
will not continue loading and fullscreen can be allowed.

webStateDidStopLoading: enables fullscreen and notifies tab model that
tab was changes. This allows removing deprecated
|webCancelStartLoadingRequest| and |webLoadCancelled:|.

BUG=674991

Review-Url: https://codereview.chromium.org/2736383006
Cr-Original-Commit-Position: refs/heads/master@{#457162}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4cc620ab46c8194efedfd1d78e7159d5968d692b