Fix Find in Page on iOS 10.3.
In iOS 10.3, an error is generated when trying to find in page on a page
with iframe:

SecurityError (DOM Exception 18): Blocked a frame with origin "https://w
ww.reddit.com" from accessing a frame with origin "https://www.redditmed
ia.com". Protocols, domains, and ports must match.

As a result, the search results are not highlighted on the page.

The fix is wrapping the offending line in try/catch.

BUG=702566
TEST=On iOS 10.3 device, open cnn.com, select tools>Find in Page" and
type any string that is present on the page. The string should be found
and highglighted, and the number of results should be correct, and
next/previous buttons should scroll the page to the next/prev result.

Review-Url: https://codereview.chromium.org/2755123002
Cr-Original-Commit-Position: refs/heads/master@{#457845}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2991f43a94bbdb2eb798a8f256b67b5b3865f129