Add disabled_test_macro.h import to showcase base test case.
Remove import from test that needs to disable a test. This makes
it easier for tests to be disabled in teh future, not requiring an
import statment to be added. This code is often never removed, so
it will make maintenance easier when tests are re-enabled.

BUG=70004

Review-Url: https://codereview.chromium.org/2761563002
Cr-Original-Commit-Position: refs/heads/master@{#458123}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb38b3e1e8d6ebfb2d2fcedcd1acb5be70f65cd9