Add disabled_test_macro import to EarlGrey base test cases.
This header is often required temporarily by tests, while a test
is disabled. The header is often not cleaned up when the test is
enabled. Adding this to the base class makes it easier to disable
tests, and keeps the test headers cleaner.

BUG=700004

Review-Url: https://codereview.chromium.org/2761443005
Cr-Original-Commit-Position: refs/heads/master@{#458132}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31a6250bc9cd79af047c0650a9f249d7e9ab2d09