Correctly reload URLs in web views if a native controller is displayed.
If a page has failed to load, a native controller is used to display
the error page.  However, if NavigationManager::Reload() is called while
an error page is displayed, we want to attempt to reload the current
NavigationItem's URL, not the native error page.

BUG=703707

Review-Url: https://codereview.chromium.org/2765903003
Cr-Original-Commit-Position: refs/heads/master@{#458499}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2d163839906a68671aea3291a7606babbfc72999