[ios] Cleanup Tab serialisation.
Remove serialisation of Tab -lastVisitedTimestamp as the value is
reset when Tab is unserialised (due to -updateLastVisitedTimestamp
call in Tab initializer).

Cache the value of Tab -tabId as it is constant during the lifetime
of the application (minor optimisation but this property is accessed
quite frequently).

Unconditionally register fullScreenController_ as a CRWWebController
observer after creation since the creation cannot fail (application
will crash if the memory allocation fail and the initializer cannot
fail under normal operation).

BUG=None

Review-Url: https://codereview.chromium.org/2769163002
Cr-Original-Commit-Position: refs/heads/master@{#459051}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 702773295d84656c8ad92eb7e700419214fbaf86