[sync] WebContentsObserver based sessions notifications
Aggregate individual WebContentsObservers to notify sync of sessions changes

This CL introduces the SyncSessionsWebContentsRouter and the
SyncSessionsRouterTabHelper, which replace the old
NotificationServiceSessionsRouter.  The SyncSessionsWebContentsRouter
aggregates events from the individual TabHelpers, which are scoped to
individual tabs, and pushes them to SessionsSyncManager.
SyncSessionsWebContentsRouters are given a source tab id property which
is set when a tab is created directly from another tab, e.g. by ctrl
clicking.
Some changes are made to SessionsHelper so that it can continue to
correctly wait for tabs to load without using Notifications.

R=zea@chromium.org
BUG=695241

Review-Url: https://codereview.chromium.org/2753753005
Cr-Original-Commit-Position: refs/heads/master@{#459225}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1901afa5e03bad9ce8a72691f746a685038241b0