[ios clean] Creates ToolsMenu Mediator and Consumer
The CL primary purpose is to create the necessary Consumer, Mediator and
ToolsMenuItem files, and hook them together.

The consumer protocol method is not intended to be final and will remain as a
placeholder for now. The model is also hardcoded and will be replaced in a future
CL.

BUG=682880

Review-Url: https://codereview.chromium.org/2769963007
Cr-Original-Commit-Position: refs/heads/master@{#460258}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8072f5360ed41e1e41265a1d528047aa4dcd3a0