[Sync] Stop accessing BrowserContextKeyedServiceFactory on non-UI thread.
In the past the AsyncDirectoryTypeController was passing a raw pointer
to a ChromeSyncClient to the model thread, where it had
GetSyncableServiceForType called on it. This was needed for several
model types, which has specific logic around only accessing their
SyncableService on their model threads. However, this was also bad, as
BrowserContextKeyedServiceFactorys were being used with a Profile
object on a non-UI thread, which is against the rules.

To fix this, GetSyncableServiceForType returns a callback which should
only be run on the model thread, that will return the same
WeakPtr<SyncableService> that we were used to. The various non-UI
thread model types do slightly different things as each type has
different rules and threading guarantees.

The most involved model type in this CL was extensions, which did not
have sufficient threading constraints to ensure safety. Added weak
pointer support to SyncValueStoreCache to fix this.

BUG=701326

Review-Url: https://codereview.chromium.org/2769113002
Cr-Original-Commit-Position: refs/heads/master@{#460304}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd6b31d4b9613087e1a68c620af83b6fe1892bdf