Move TestingPrefService to use unique_ptr<Value>
Various Set*Pref methods of TestingPrefService pass Value ownership with a raw
pointer.

This CL changes that to passing unique_ptr, because that is a safer and clearer
option. It is also a pre-requisite for migrating Value API off raw pointer
ownership passing.

BUG=697817

Review-Url: https://codereview.chromium.org/2782553004
Cr-Original-Commit-Position: refs/heads/master@{#460363}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8684c9acc69aa307f5fabfef0182b61c7fe8ce00