Reload with empty navigation manager should not crash
Previously, the app crashes when user tries to reload when navigation
manager is empty.

This CL fixes the issue by adding an earl return to |Reload| in
NavigationManagerImpl so that it does nothing when its empty.

This CL also adds corresponding unit tests to prevent regressions.

BUG=705707

Review-Url: https://codereview.chromium.org/2780703002
Cr-Original-Commit-Position: refs/heads/master@{#460464}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e70e89bec8b805f1056ba85e1e14c9486341d138