Use correct snapshot for the tab transition animation on iPad.
The TabSwitcherTransitionContextContent stores the index of the tab from
which the user animated in. However the tabs can be manipulated in the
switcher (added removed) and so the tab at index i will not necessarily
be the same as the tab that was originally at index i when first
entering the switcher.

This can lead to an incorrect snapshot being used.

To reproduce:

1. Have a clean setup, just a single tab.
2. Navigate to a website which looks different to Google.com
e.g. cnn.com
3. Enter the tab switcher (there should only be one tab)
4. Close the tab in the switcher
5. Hit the plus (+) button to create a new tab
6. The new tab will animate to full screen with the old (cnn.com)
snapshot

This change removes the use of an index to track the validity of the
tab snapshots. Instead using the tabID of the Tab
object itself to see if it's different.

BUG=706891

TEST=Have a clean setup, just a single tab. Navigate to (e.g.) cnn.com.
Enter the tab switcher. Close the tab in the switcher. Hit the plus (+)
button to create a new tab. The new tab will animate to full screen with
a correct snapshot.

Review-Url: https://codereview.chromium.org/2776083003
Cr-Original-Commit-Position: refs/heads/master@{#460905}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6d1b157721d10aebaca323f514ba7d5a09bdcef