Fixes History page keeps on loading when signed in with Passphrase enabled account.
This bug was exposed as a result of https://codereview.chromium.org/2589313002.
There are two types of data types for sync:
Preferred type - The sync service will try to make active exactly these types.
Active: a preferred datatype that is actively being synchronized.
On iOS sync service, when fetching the current state of a data type, we were
 looking up preferred data type, instead of active. Hence even though the
 passphrase is enabled, and HISTORY_DELETE_DIRECTIVES is disabled in this
stage, it was being returned as true, since it was still a part of the preferred
data types.

TEST=Test with an account with secondary passphrase. The history view must
 only load local history when passphrase is enabled. Otherwise load synced
 history from across devices enabled

BUG=679347
TBR=skym@chromium.org

Review-Url: https://codereview.chromium.org/2790483002
Cr-Original-Commit-Position: refs/heads/master@{#461642}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54106f055fcaed85a4cfd34d749eaf7ebf107137