Created SessionCertificatePolicyCache.
This is a public web// interface for an object owned by WebStates.
It holds the certificates allowed for a specific WebState.
CRWSessionCertificatePolicyManager's serialization code has been
moved to CRWSessionCertificatePolicyCacheStorage.

This approach is analogous to content's HostContentSettingsMap,
a refcounted keyed service that's persisted with the Profile.

This CL also introduces the ios_web_unittests_arc target.
BUG=454984

Review-Url: https://codereview.chromium.org/2769573003
Cr-Original-Commit-Position: refs/heads/master@{#461941}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d3a3986fe2f2e5498807e2cfe23865a0d00240d2