[ios] Extend BrowserList API.
Add a method BrowserList::GetIndexOfBrowser() to find the index of
a Browser instance to simplify closing Browsers (since Browser are
closed by index, not pointer).

Rename GetBrowserCount() method to count() and inline it as it is
a simple function (just a wrapper around std::vector::size()) and
not inlining it could make the iteration over the BrowserList slow.

BUG=None

Review-Url: https://codereview.chromium.org/2798863004
Cr-Original-Commit-Position: refs/heads/master@{#462452}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3a375fc9d871fe97a5a2bc23e5471d95619f1c49