Fixing testBrowsingPostEntryWithButton flakiness.
This test has been flaky on devices because Earl Grey fails to
synchronize with omnibox text.

This CL first fixes the flakiness by adding a wait to wait for omnibox
text, and then this test seems to be more related to form submission,
so moving it to FormsTestCase.

BUG=

Review-Url: https://codereview.chromium.org/2806743004
Cr-Original-Commit-Position: refs/heads/master@{#463415}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33c5ac234dc93a3cfc31fda63fe716d8d2a40410