Removes wifi dependency for testHistoryEntriesStatusCell, so that it passes reliably when running on devices with WIFI turned off.
The test, was checking the web contents of the URL as the last step
which is unnecessary. This test is meant to test the history entries,
and that the right number of tabs are opened in the end. The
contents of the URL are not important.

BUG=709135

Review-Url: https://codereview.chromium.org/2806253002
Cr-Original-Commit-Position: refs/heads/master@{#463441}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ac504ddefaadc3b933db8cffde4a599539c89b79