Fixing UI glitch in settings view.
When clicking on the "CONTINUE AS xxx" to sign-in, SigninInteractionController
presents a signed-in view and sign-in the user at the same time. When the user
is signed in, the settings collection view is reloaded to update the sign-in
cell (while the signed-in view is sliding up).
This situation creates a UI glitch.

To avoid a double animations, SettingsCollectionViewController should not reload
the collection view when the controller is overlaid. The sign-in item should
be reloaded when -[SettingsCollectionViewController viewWillAppear] is called.

BUG=661794

Review-Url: https://codereview.chromium.org/2809003005
Cr-Commit-Position: refs/heads/master@{#463762}
Committed: https://chromium.googlesource.com/chromium/src/+/725f9b03d56b428dffc261c95816e17904f93b53
(cherry picked from commit 725f9b03d56b428dffc261c95816e17904f93b53)

Review-Url: https://codereview.chromium.org/2813263003
Cr-Original-Commit-Position: refs/heads/master@{#464709}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 697a0aa44a8ea1bbe4e587845856ef0ccb687e7d