Convert main NTP panel to UIViewController.
To ease modernizing the NTP only the internal panels are being converted
to UIViewControllers.  This means all the plumbing between the
BrowserViewController and the internal NTP panels (WebController, NTP)
hierarchy is skipped.  While normally the logic to push and pop a view
controller would be owned by a coordinator, in this case the old NTP
controller adds and removes child view controllers itself when a load
is initiated, and when WebController calls -willBeDismissed.

BUG=708319

Review-Url: https://codereview.chromium.org/2806153004
Cr-Original-Commit-Position: refs/heads/master@{#464850}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f6e1cbafd32c584fe6197eaefd4edd2f08e1024