Re-land "Replace requirePageReconstruction with setCustomUserAgent"
This CL re-lands the "Replace requirePageReconstruction with
setCustomUserAgent" CL.

The following is the original CL descriptions:

WKWebView provides a property: customUserAgent that we can use to
change the user agent associated with the web view, which means that we
don't have to requirePageConstruction anymore when user agent changes.

After applying this change, I observed that it's becoming significantly
faster to navigate between desktop and mobile pages during back and
forward navigations.

BUG=707368

Review-Url: https://codereview.chromium.org/2822003002
Cr-Original-Commit-Position: refs/heads/master@{#465042}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1688d0107c311a09b0bfb03df0b81ab9b21afc0