[ios] Mark the FIP UI as inactive before disabling highlighting.
The handler that BVC passes to |disableFindInPageWithCompletionHandler:| checks
the value of FindTabHelper::IsFindUIActive() and shows or hides the find bar
accordingly.  Since FindTabHelper::StopFinding() is attempting to disable find in
page, we need to set the UI to inactive before the completion handler runs.

BUG=709475

Review-Url: https://codereview.chromium.org/2817673002
Cr-Original-Commit-Position: refs/heads/master@{#465350}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 578c2d76b73852af45e3d419676bd2e0c0b1b59a