Display the web view after commit for non-user-initiated loads.
The toolbar's visibility and URL is dependent on the visible URL.
However, the previous implementation of |-ensureWebViewCreated| would
immediately add web view to the hierarchy, even for non-user-initiated
loads.  This CL updates the web view display logic so that it is
displayed at the same time as the toolbar updating.

BUG=703222

Review-Url: https://codereview.chromium.org/2820013003
Cr-Original-Commit-Position: refs/heads/master@{#465365}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cd2e9b497d8623d1c2844b6df874e42f865c2631