Remove URLRequestJob::GetResponseCode implementations outside of net/.
URLRequestJob itself now has a default implementation, which gets
the response code from the headers returned by GetResponseInfo.
We eventually want to remove this method, since it's redundant,
and has weird behavior for just when it is and is not safe to
call.

BUG=688481

Review-Url: https://codereview.chromium.org/2827233005
Cr-Original-Commit-Position: refs/heads/master@{#466631}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ec95f5b430aac7baff41e4b34cda43116a438905