Remove the use of custom logic to determine search suggestion types.
iOS has a custom method for determining which omnibox suggestion types
are search/query types (as opposed to URL types). This change removes it
in favor of the cross-platform method used on other platforms. This has
useful effects on how we display navsuggest and calculator suggestions
and eliminates the need to maintain separate logic.

See the bug for a detailed description of the motivation for this change
and an analysis of its effects.

BUG=716047

Review-Url: https://codereview.chromium.org/2852553002
Cr-Original-Commit-Position: refs/heads/master@{#468317}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 97567bc01417a53c3157e0fe0fb3e8c4ce714db1